requirements.txt
fastapi
uvicorn
astrapy
python-dotenv
argon2-cffi
python-jose
apscheduler
pydantic[email]
pydantic
email-validator